backend/requirements.txt
# MedNorm AI — Backend Dependencies
# HackMatrix 2.0 · Team LegacyCoderz — Track 2

# ── Core Framework ─────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic[email]==2.9.2
python-multipart==0.0.12

# ── PDF Parsing ────────────────────────────────────────────
pdfplumber==0.11.4          # PDF text + table extraction
pypdf==5.0.0               # PDF metadata / page count

# ── OCR (for scanned documents) ───────────────────────────
Pillow==11.0.0             # Image processing
pytesseract==0.3.13        # Tesseract OCR wrapper
# Note: Tesseract binary must be installed separately:
#   sudo apt install tesseract-ocr tesseract-ocr-hin  (Ubuntu/Debian)
#   brew install tesseract   (macOS)

# ── Clinical NLP / AI ─────────────────────────────────────
openai==1.54.0              # GPT-4 for clinical entity extraction

# ── FHIR ──────────────────────────────────────────────────
fhir.resources==7.1.0      # FHIR R4 resource validation (optional)

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.1       # Load .env for API keys
httpx==0.27.2              # Async HTTP client
orjson==3.10.7             # Fast JSON serialization

# ── Dev / Testing ──────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx                      # Also used for TestClient
